vhost: add support for negotiating extended features.
Similar to virtio infra, vhost core maintain the features
status in the widest format available and allow the devices
to implement extended version of the getter/setter.

Some care is needed for features bit manipulation: when clearing
a bit with 'and not' bitwise operations, the bit mask must be
extended to the feature format, or the all highest bits will be
unintentionally cleared.

Note that 'protocol_features' are not extended: they are only
used by vhost-user, and the latter device is not going to implement
extended features soon.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>